Remove backend usage from abstract callbacks
